archive/crm_modules/Estructura de carpetas/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
pyotp
requests
paypalrestsdk
jinja2
python-multipart
email-validator
cryptography
